src/beam/requirements.txt
# Apache Beam Pipeline Requirements
# Specific versions for Dataflow compatibility

apache-beam[gcp]==2.54.0
google-cloud-bigquery==3.17.1
google-cloud-storage==2.10.0

# Data Processing
pandas==2.1.4
pyarrow==15.0.0
numpy==1.26.3

# Data Quality
great-expectations==0.18.4

# APIs
requests==2.31.0

# Logging
python-json-logger==2.0.7
structlog==24.1.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-timeout==2.2.0

# Development
ipython==8.18.1
jupyter==1.0.0
